Fix global center of mass for `Forces` and add `PhysicsTransformHelper` (#787)
# Objective

`Forces` has methods like `apply_force_at_point` that use the `GlobalCenterOfMass`. However, the `GlobalCenterOfMass` can be incorrect in two common cases:

1. At spawn, it is currently not initialized correctly until the first time physics runs.
2. If you change the `Transform` (e.g. to teleport an entity), it is only updated the next time physics runs.

This can lead to wildly incorrect torques or angular impulses as the force is applied relative to an incorrect center of mass.

Problem 1 is easy to fix with an observer, but problem 2 is very difficult. We cannot know the up-to-date global physics position without first running some form of transform propagation, and we cannot immediately react to `Transform` changes either.

For now, I think we should simply provide APIs for people to update global physics positions in an ad hoc way, and to remove the reliance on `GlobalCenterOfMass` in favor of simply computing it based on the `Position`, `Rotation`, and `ComputedCenterOfMass`.

## Solution

Add a `PhysicsTransformHelper` system parameter for updating the physics transform of a given entity, and remove `GlobalCenterOfMass`.

Additionally, I fixed some problems with `GlobalTransform`, `ColliderTransform`, and mass properties being incorrect until the first time physics runs. They should now be correct right after spawn.