Fix UB and improve parallel iteration in solver (#790)
# Objective

The solver currently has some UB that is making local determinism tests fail on Windows. This is because we use parallel iteration even for the overflow constraints that should be solved serially.

## Solution

Solve overflow serially. This will be reworked further when we implement wide SIMD.

I also changed the `par_for_each!` macro to a normal function, which fixes autocomplete when using the macro, and added some bounds on when to use parallel iteration and when to use serial iteration.