Fix colliders not working after removing `Disabled`/`ColliderDisabled` (#791)
# Objective

Fixes #786.

Disabling colliders is currently bugged. After removing `Disabled`/`ColliderDisabled`, they do not work again. This is because the queries in the observers responsible for re-enabling are incorrectly filtering out disabled entities.

## Solution

Fix the observers. There's some nasty duplication since the queries have to be subtly different, but it can be cleaned up later.