Expose anchors, world-space `point`, and `normal_speed` for contacts (#792)
# Objective

Currently, `ContacttPoint` stores a `local_point1` and `local_point2` on the surfaces of the two shapes. These are the closest points in local space.

However:

- The solver needs the world-space contact points relative to the centers of mass. The local points are not actually needed for anything, they are just the output of Parry.
- The solver should not be using the closest points for constraints. There should be a single world-space contact point that is the midpoint between the two closest points, and world-space anchors that correspond to this point.
- Users typically only need world-space values, but getting them currently requires manual transformations.
- For users, it is confusing that a `ClosestPoint` type actually stores two points, one for each shape.

We can do better! We can store the world-space anchors directly in `ContactPoint`, and also store a single world-space `point` for debugging and user convenience.

## Solution

Remove `local_point1` and `local_point2` in favor of a world-space `anchor1` and `anchor2` relative to the center of mass. These are computed in the narrow phase, removing the need to query for collider data in `prepare_contact_constraints`.

A world-space `point` for the midpoint between the closest points is now stored and used for debug rendering. The contact anchors point to this.

Additionally, I have exposed a `normal_speed` property for `ContactPoint`, computed by `ContactConstraint`.
It is computed before the solver, and can be used as an impact velocity to determine how "strong" the contact is in a mass-independent way. `-normal_speed` is equivalent to how fast the shapes are approaching each other at this point.

Unrelated to the core changes, I also fixed the linear part of the effective mass computation not taking locked axes into account.

---

## Migration Guide

The `local_point1` and `local_point2` properties of `ContactPoint` have been removed in favor of a world-space `anchor1` and `anchor2` relative to the center of mass.

The `global_point1` and `global_point2` methods have also been removed, but a new world-space `point` is available for the midpoint between the closest points.