Improve inverse angular inertia tensor computation (#793)
# Objective

Closes #629.

`ComputedAngularInertia` updates in 3D currently invert the tensor rather than inverting the principal moments of inertia and using those to construct the matrix. This is both less efficient and has a more limited range of supported inertia values, as the determinant gets close to zero.

## Solution

Invert the principal angular inertia vector instead. Testing with a cuboid, this brings the range of supported mass values from about about `1e13` to about `1e19`.