Contact pruning for 3D (#798)
# Objective

Contact manifolds in 3D should be limited to 4 contact points. This is important for both stability and performance, especially when we implement wide SIMD contacts.

Some specific cases (some convex polyhedron collisions or possible custom shapes) can produce more than 4 points. We should have a contact pruning algorithm to get rid of these extra points.

## Solution

Implement a contact pruning algorithm that chooses the optimal points based on the distance to the center of mass, the penetration depth, and the contact area. The implementation is mostly based on Jolt.

In the future, we can combine this with a manifold reduction scheme that combines manifolds that have similar normals, pruning the resulting larger manifold.