Add `JointGraph` and `JointCollisionDisabled` (#799)
# Objective

Resolves #259.
Closes #274.

A common requirement when using joints is to disable collisions between the attached bodies, for example a wheel and a car. Collision layers are typically not suitable for this, as you may still want a car to collide with the wheels of *other* cars. Implementing the desired behavior properly requires custom hooks, which is both inefficient and overly complicated for such a common scenario.

We should have a first-party `JointCollisionDisabled` component. To implement it efficiently, we also need to maintain an internal `JointGraph` to track which joints are connected to which bodies. A joint graph is also needed for #578.

## Solution

- Implement a `JointGraph` that is managed for each joint with a `JointGraphPlugin<T: Joint>`. To make it work properly, a single entity can now only hold up to one type of joint component. If a joint is added on an entity that already has a joint, the old joint is automatically removed, and a warning is emitted if `debug_assertions` are enabled.
- Add a `JointCollisionDisabled` component for disabling collisions between bodies attached by joints.

---

## Migration Guide

Entities can now only hold up to one type of joint component. If a joint is added on an entity that already has a joint, the old joint is automatically removed, and a warning is emitted if `debug_assertions` are enabled. A rigid body can still be attached to other bodies by multiple joints, but each joint must be on its own entity.

`ContactEdge` now also stores the rigid body entities.