Allocate free IDs lowest-first using an `IdPool` for `StableUnGraph` (#807)
# Objective

Contact pairs maintain stable indices using the `StableUnGraph`. It marks removes nodes and edges and vacant, and uses linked lists for free node and edge lists.

This results in reusing nodes and edges in removal order. The problem is that it causes perceived determinism issues when contacts are removed: when "restarting" a scene, old freed contacts affect the order of new reused contact edges, often causing different results. This can be seen in the `determinism_2d` example; respawning the scene produces different results *unless* the contact graph is cleared.

Ideally, this should not be something users need to worry about. Respawning the same scene should produce the same results each time without clearing any caches. This PR aims to fix this by allocating freed contact IDs in a lowest-first order, not in removal order.

## Solution

Add an `IdPool` that stores freed `u32` IDs in a min-heap, and use it for the free lists in the `StableUnGraph`. This way, the lowest IDs are allocated first.

Running the `determinism_2d` example, it is no longer necessary to manually clear any cached contact data to get deterministic results when resetting the scene.