fix temporaries freed while still in use (#819)
fix "temporaries freed while still in use" across a couple of function uses.

## Solution

I took the approach of keeping the clones of things like `DebugName` from the `.name` function, etc and instead stored the value so it could be referenced, since the usage relies on deref coercion to go from `DebugName` to `&str`.

The examples run after this fix