Improve tunneling prevention by solving static contacts last (#824)
# Objective

Currently, static contacts are mixed with dynamic contacts starting from the second graph color. To improve tunneling prevention, it'd be better to reserve some colors just for static contacts, and to solve them last.

## Solution

- Increase the `GRAPH_COLOR_COUNT` from 12 to 24
- Reserve some colors just for static contacts, and build static colors from the end to give them higher priority

In the future, we could consider using dedicated one-body contact constraints for the static colors to further optimize things.