Fix `glam_matrix_extensions` and `bevy_heavy` commit refs (#832)
# Objective

The `main` branch is currently broken, because it depends on the main branches of `glam_matrix_extensions` and `bevy_heavy` without specifying a git ref, and they have already been updated to Bevy 0.17. We need to specify a git ref that uses the old versions.

## Solution

Specify functional git refs for `glam_matrix_extensions` and `bevy_heavy` .