Allow adding/removing `Sleeping` to sleep/wake bodies (#834)
# Objective

In past versions of Avian, adding/removing `Sleeping` manually was supported as a way to sleep or wake up a rigid body. However, after #809, this does not work properly, and you should instead use the `SleepBody` and `WakeBody` commands.

Additionally, `SleepBody` does not currently split the island, which can cause bodies that are not in contact to also be forced asleep.

## Solution

Split islands when using `SleepBody`, and add hooks to automatically sleep/wake bodies when `Sleeping` is manually added/removed.