add ApplyPosToTransform component to include an entity in  sync (#837)
# Objective

There are situations (networking) where the Position/Rotation components are networked, but not RigidBody. For example for other players' entities I only network Position/Rotation but not physics components since I don't need to simulate their behaviour.
In those cases it would be useful to have a way to run the `position_to_transform` system on these entities.


## Solution

Add `ApplyPosToTransform` component for this purpose.