Add PartialEq bound on Sleeping
I would like to also enable rollback for the Sleeping component, so that on any rollback we restore Sleeping to what it was at the time of rollback. 

My `add_rollback` has a `PartialEq` bound because usually I need to compare the current value of the component with the past value, so it would be convenient for me if Sleeping also had this bound.