Disable mass property warning for collider constructors (#849)
# Objective

Rigid bodies with a `ColliderConstructor` or `ColliderConstructorHierarchy` often emit a warning for invalid mass properties before the collider has actually been generated. Annoying!

## Solution

Skip the warning for rigid bodies with a `ColliderConstructor` or `ColliderConstructorHierarchy`. Once the colliders have been generated, the mass is changed, which will automatically check for invalid mass again.